move quota-related functions to a separate file
